abr-shell-ui/src/main/kotlin/com/d20charactersheet/adventurebookresolver/shellui/command/DeleteCommand.kt
package com.d20charactersheet.adventurebookresolver.shellui.command

import com.d20charactersheet.adventurebookresolver.shellui.domain.AdventureBookResolver
import com.d20charactersheet.adventurebookresolver.shellui.services.ConsoleService
import com.d20charactersheet.adventurebookresolver.shellui.services.displayActions
import org.springframework.shell.standard.ShellComponent
import org.springframework.shell.standard.ShellMethod

@ShellComponent
class DeleteCommand(val adventureBookResolver: AdventureBookResolver, val consoleService: ConsoleService) {

    @ShellMethod("delete action from current entry to given entry id")
    fun delete(entryId: Int) {
        with(adventureBookResolver.book) {
            delete(entryId)
            displayActions(consoleService)
        }
    }

}
